Move PaxConfig to its own file
