+++ Pax.csproj
@@ -24,6 +24,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AssemblyInfo.cs" />
+    <Compile Include="PaxConfig.cs" />
     <Compile Include="Paxifax.cs" />
     <Compile Include="Pax.cs" />
   </ItemGroup>
